# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Matching Apps API",
        "description": "Implementation of Swagger in Laravel with Passport Auth Documentation",
        "contact": {
            "email": "mark@mosaique.link"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user/{id}/update": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Update",
                "description": "{id} is the User ID from users table",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "mobile_number",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "phone_number",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "date_of_birth",
                        "in": "query",
                        "schema": {
                            "type": "date"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "address",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "about_you",
                        "in": "query",
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "fields_of_interest",
                        "in": "query",
                        "description": "Array of Interest and use json_encode",
                        "schema": {
                            "type": "json"
                        }
                    },
                    {
                        "name": "skills",
                        "in": "query",
                        "description": "Array of Skills and use json_encode",
                        "schema": {
                            "type": "json"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Array of Language and use json_encode",
                        "schema": {
                            "type": "json"
                        }
                    },
                    {
                        "name": "qualification",
                        "in": "query",
                        "description": "Array of Qualification and use json_encode",
                        "schema": {
                            "type": "json"
                        }
                    },
                    {
                        "name": "other",
                        "in": "query",
                        "schema": {
                            "type": "text"
                        }
                    },
                    {
                        "name": "avatar",
                        "in": "query",
                        "description": "this should be image upload",
                        "schema": {
                            "type": "image"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "https://matching-apps.test/oauth/authorize",
                        "tokenUrl": "https://matching-apps.test/oauth/token",
                        "refreshUrl": "https://matching-apps.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}